requirements.txt
Flask==3.0.0
Pillow==10.1.0
pillow-heif==0.13.1
RPi.GPIO==0.7.1
spidev==3.6
Werkzeug==3.0.1